fix elint on book.js
