+++ src/redux/Store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from './book/bookSlice.js';
+import booksReducer from './books/bookSlice.js';
 import categoriesReducer from './categories/categoriesSlice.js';
 
 const store = configureStore({
